Breaks model module down into more manageable submodules
