+++ snat_sim/pipeline/pipelines.py
@@ -10,8 +10,7 @@ from egon.pipeline import Pipeline
 from .lc_fitting import FitLightCurves, FitResultsToDisk
 from .lc_simulation import SimulateLightCurves, SimulationToDisk
 from .plasticc_io import LoadPlasticcSims
-from ..models import SNModel
-from ..reference_stars import VariableCatalog
+from ..modeling import SNModel, VariableCatalog
 
 
 class FittingPipeline(Pipeline):
