+++ tests/test_modeling/test_pwv/testStaticPWVTrans.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from snat_sim import models
+from snat_sim.modeling import pwv
 from .base import PropagationEffectTests
 
 
@@ -12,19 +12,19 @@ class BaseTests(PropagationEffectTests, TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.propagation_effect = models.StaticPWVTrans()
+        cls.propagation_effect = pwv.StaticPWVTrans()
 
     def test_propagation_includes_pwv_transmission(self):
         """Test the ``propagate`` method applies PWV absorption"""
 
         # Get the expected transmission
-        pwv = 5
+        pwv_concentration = 5
         wave = np.arange(4000, 5000)
-        transmission_model = models.FixedResTransmission(resolution=self.propagation_effect._transmission_res)
-        transmission = transmission_model.calc_transmission(pwv=pwv, wave=wave)
+        transmission_model = pwv.PWVTransmissionModel(resolution=self.propagation_effect._transmission_res)
+        transmission = transmission_model.calc_transmission(pwv=pwv_concentration, wave=wave)
 
         # Get the returned flux
-        self.propagation_effect._parameters = [pwv]
+        self.propagation_effect._parameters = [pwv_concentration]
         propagated_flux = self.propagation_effect.propagate(wave, np.ones_like(wave))
         np.testing.assert_equal(transmission, propagated_flux[0])
 
@@ -34,7 +34,7 @@ class DefaultParameterValues(TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.propagation_effect = models.StaticPWVTrans()
+        cls.propagation_effect = pwv.StaticPWVTrans()
 
     def test_default_pwv_is_zero(self):
         """Test the default ``pwv`` parameter is 0"""
