+++ tests/test_modeling/test_supernovae/testObservedCadence.py
@@ -4,8 +4,8 @@ from unittest import TestCase
 
 import numpy as np
 
-from snat_sim.models import ObservedCadence
-from ..mock import create_mock_plasticc_light_curve
+from snat_sim.modeling.supernova import ObservedCadence
+from ...mock import create_mock_plasticc_light_curve
 
 
 class SncosmoFormatting(TestCase):
