+++ tests/test_pipeline/test_lc_simulation/testSimulateLightCurves.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 import numpy as np
 from egon.mock import MockSource, MockTarget
 
-from snat_sim.models import ObservedCadence, SNModel
+from snat_sim.modeling import ObservedCadence, SNModel
 from snat_sim.pipeline.lc_simulation import SimulateLightCurves
 from tests.mock import create_mock_plasticc_light_curve
